Refactor user flags provider to unified provider
