+++ WordPress/src/main/java/org/wordpress/android/localcontentmigration/LocalContentEntity.kt
@@ -5,6 +5,7 @@ import org.wordpress.android.fluxc.model.PostModel
 enum class LocalContentEntity(private val isSiteContent: Boolean = false) {
     EligibilityStatus,
     AccessToken,
+    UserFlags,
     Site,
     Post(isSiteContent = true),
     ;
@@ -27,6 +28,7 @@ sealed class LocalContentEntityData {
     ): LocalContentEntityData()
 
     data class AccessTokenData(val token: String): LocalContentEntityData()
+    data class UserFlagsData(val flags: Map<String, Any?>): LocalContentEntityData()
     data class SitesData(val localIds: List<Int>): LocalContentEntityData()
     data class PostsData(val localIds: List<Int>): LocalContentEntityData()
     data class PostData(val post: PostModel) : LocalContentEntityData()
