+++ WordPress/src/main/java/org/wordpress/android/userflags/resolver/UserFlagsResolver.kt
@@ -1,25 +1,19 @@
 package org.wordpress.android.userflags.resolver
 
-import android.database.Cursor
-import org.wordpress.android.provider.query.QueryResult
-import org.wordpress.android.resolver.ContentResolverWrapper
+import org.wordpress.android.localcontentmigration.LocalContentEntity.UserFlags
+import org.wordpress.android.localcontentmigration.LocalContentEntityData.UserFlagsData
+import org.wordpress.android.localcontentmigration.LocalMigrationContentResolver
 import org.wordpress.android.ui.prefs.AppPrefsWrapper
 import org.wordpress.android.userflags.JetpackLocalUserFlagsFlag
 import org.wordpress.android.userflags.UserFlagsAnalyticsTracker
 import org.wordpress.android.userflags.UserFlagsAnalyticsTracker.ErrorType
-import org.wordpress.android.userflags.provider.UserFlagsProvider
-import org.wordpress.android.util.publicdata.WordPressPublicData
-import org.wordpress.android.viewmodel.ContextProvider
 import javax.inject.Inject
 
 class UserFlagsResolver @Inject constructor(
     private val jetpackLocalUserFlagsFlag: JetpackLocalUserFlagsFlag,
-    private val contextProvider: ContextProvider,
-    private val wordPressPublicData: WordPressPublicData,
-    private val queryResult: QueryResult,
-    private val contentResolverWrapper: ContentResolverWrapper,
     private val appPrefsWrapper: AppPrefsWrapper,
-    private val userFlagsAnalyticsTracker: UserFlagsAnalyticsTracker
+    private val userFlagsAnalyticsTracker: UserFlagsAnalyticsTracker,
+    private val localMigrationContentResolver: LocalMigrationContentResolver,
 ) {
     fun tryGetUserFlags(onSuccess: () -> Unit, onFailure: () -> Unit) {
         val isFeatureFlagEnabled = jetpackLocalUserFlagsFlag.isEnabled()
@@ -34,35 +28,21 @@ class UserFlagsResolver @Inject constructor(
         }
         userFlagsAnalyticsTracker.trackStart()
         appPrefsWrapper.saveIsFirstTryUserFlagsJetpack(false)
-        val userFlagsResultCursor = getUserFlagsResultCursor()
-        if (userFlagsResultCursor != null) {
-            val userFlags = queryResult.getValue<Map<String, Any?>>(userFlagsResultCursor) ?: emptyMap()
-            if (userFlags.isNotEmpty()) {
-                val success = updateUserFlags(userFlags)
-                if (success) {
-                    userFlagsAnalyticsTracker.trackSuccess()
-                    onSuccess()
-                } else {
-                    userFlagsAnalyticsTracker.trackFailed(ErrorType.UpdateUserFlagsError)
-                    onFailure()
-                }
+        val (userFlags) = localMigrationContentResolver.getDataForEntityType<UserFlagsData>(UserFlags)
+        if (userFlags.isNotEmpty()) {
+            val success = updateUserFlags(userFlags)
+            if (success) {
+                userFlagsAnalyticsTracker.trackSuccess()
+                onSuccess()
             } else {
-                userFlagsAnalyticsTracker.trackFailed(ErrorType.NoUserFlagsFoundError)
+                userFlagsAnalyticsTracker.trackFailed(ErrorType.UpdateUserFlagsError)
                 onFailure()
             }
         } else {
-            userFlagsAnalyticsTracker.trackFailed(ErrorType.QueryUserFlagsError)
-            onFailure()
-        }
-    }
-
-    private fun getUserFlagsResultCursor(): Cursor? {
-        val wordpressUserFlagsUriValue =
-                "content://${wordPressPublicData.currentPackageId()}.${UserFlagsProvider::class.simpleName}"
-        return contentResolverWrapper.queryUri(
-                contextProvider.getContext().contentResolver,
-                wordpressUserFlagsUriValue
-        )
+                userFlagsAnalyticsTracker.trackFailed(ErrorType.NoUserFlagsFoundError)
+                onFailure()
+            }
+//            userFlagsAnalyticsTracker.trackFailed(ErrorType.QueryUserFlagsError)
     }
 
     @Suppress("TooGenericExceptionCaught", "SwallowedException")
