+++ WordPress/src/wordpress/AndroidManifest.xml
@@ -3,11 +3,6 @@
 
     <application>
         <provider
-            android:name=".userflags.provider.UserFlagsProvider"
-            android:authorities="${applicationId}.UserFlagsProvider"
-            android:exported="true" />
-
-        <provider
             android:name=".bloggingreminders.provider.BloggingRemindersProvider"
             android:authorities="${applicationId}.BloggingRemindersProvider"
             android:exported="true" />
